create common base class for custom esphome communication components and simplify code
