+++ components/km271_wifi/km271_params.cpp
@@ -1,88 +1,41 @@
 #include "km271_params.h"
 #include "esphome/core/log.h"
-#include "esphome/core/util.h"
 #include "esphome/core/hal.h"
 #include <stdint.h>
 #include "3964r.h"
+#include "km271_communication_components.h"
 
 namespace esphome {
 namespace KM271 {
 
 static const char *const TAG = "BPARM";
 
-BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc* paramDesc, esphome::sensor::Sensor *sensor):
+BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc *paramDesc, esphome::sensor::Sensor *sensor):
     paramDesc(paramDesc),
+    communicationComponent(nullptr),
     sensor(sensor),
-    binarySensor(nullptr),
-    switch_(nullptr),
-    number(nullptr),
-    select(nullptr),
-    assembler(nullptr)
+    binarySensor(nullptr)
 {
 
 }
 
-BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc* paramDesc, esphome::binary_sensor::BinarySensor *sensor):
+BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc *paramDesc, esphome::binary_sensor::BinarySensor *binarySensor):
     paramDesc(paramDesc),
+    communicationComponent(nullptr),
     sensor(nullptr),
-    binarySensor(sensor),
-    switch_(nullptr),
-    number(nullptr),
-    select(nullptr),
-    assembler(nullptr)
+    binarySensor(binarySensor)
 {
 
 }
 
-BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc* paramDesc, BuderusParamSwitch *switch_):
+BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc* paramDesc, CommunicationComponent * communicationComponent):
     paramDesc(paramDesc),
+    communicationComponent(communicationComponent),
     sensor(nullptr),
-    binarySensor(nullptr),
-    switch_(switch_),
-    number(nullptr),
-    select(nullptr),
-    assembler(nullptr)
+    binarySensor(nullptr)
 {
 
 }
-
-BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc* paramDesc, BuderusParamNumber *paramNumber):
-    paramDesc(paramDesc),
-    sensor(nullptr),
-    binarySensor(nullptr),
-    switch_(nullptr),
-    number(paramNumber),
-    select(nullptr),
-    assembler(nullptr)
-{
-
-}
-
-BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc *paramDesc, BuderusParamSelect *paramSelect):
-    paramDesc(paramDesc),
-    sensor(nullptr),
-    binarySensor(nullptr),
-    switch_(nullptr),
-    number(nullptr),
-    select(paramSelect),
-    assembler(nullptr)
-{
-
-}
-
-BuderusValueHandler::BuderusValueHandler(const t_Buderus_R2017_ParamDesc *paramDesc, MultiParameterUnsignedIntegerAssembler *assembler):
-    paramDesc(paramDesc),
-    sensor(nullptr),
-    binarySensor(nullptr),
-    switch_(nullptr),
-    number(nullptr),
-    select(nullptr),
-    assembler(assembler)
-
-{
-
-}
-
 
 uint32_t parseUnsignedInteger(uint8_t *data, size_t len)
 {
@@ -111,312 +64,100 @@ int32_t parseSignedInteger(uint8_t *data, size_t len)
 
 void BuderusValueHandler::parseAndTransmit(uint8_t *data, size_t len)
 {
-    if (sensor) {
-        if (paramDesc->sensorType == UNSIGNED_INT) {
-            uint32_t value = parseUnsignedInteger(data, len);
-            sensor->publish_state(value);
-        } else if (paramDesc->sensorType == SIGNED_INT) {
-            int32_t value = parseSignedInteger(data, len);
-            sensor->publish_state(value);
-        } else if (paramDesc->sensorType == UNSIGNED_INT_DIVIDED_BY_2) {
-            uint32_t value = parseUnsignedInteger(data, len);
-            sensor->publish_state(((float)value) / 2.0f);
-        } else {
-            ESP_LOGW(TAG, "Sensor type %d NYI", paramDesc->sensorType);
-        }
-    } else if(binarySensor) {
-        if (paramDesc->sensorType == UNSIGNED_INT) {
-            binarySensor->publish_state(data[len-1] > 0);
-        } else if(paramDesc->sensorType == BIT_AT_OFFSET) {
-            uint16_t bitOffset = paramDesc->sensorTypeParam;
-            uint16_t byteOffset = 0;
-            while (bitOffset >= 8) {
-                byteOffset += 1;
-                bitOffset -= 8;
-                if (byteOffset >= len) {
-                    ESP_LOGW(TAG, "Parameter 0x%04x has bit offset %d but message only has %d bytes", paramDesc->parameterId, paramDesc->sensorType, len);
-                    return;
-                }
+    if (paramDesc->sensorType == UNSIGNED_INT) {
+        uint32_t value = parseUnsignedInteger(data, len);
+        handleReceivedUnsignedValue(paramDesc->sensorTypeParam, value);
+    } else if (paramDesc->sensorType == SIGNED_INT) {
+        int32_t value = parseSignedInteger(data, len);
+        handleReceivedUnsignedValue(paramDesc->sensorTypeParam, value);
+    } else if (paramDesc->sensorType == UNSIGNED_INT_DIVIDED_BY_2) {
+        uint32_t value = parseUnsignedInteger(data, len);
+        handleReceivedFloatValue(paramDesc->sensorTypeParam, (float)value / 2.0f);
+    } else if(paramDesc->sensorType == BIT_AT_OFFSET) {
+        uint16_t bitOffset = paramDesc->sensorTypeParam;
+        uint16_t byteOffset = 0;
+        while (bitOffset >= 8) {
+            byteOffset += 1;
+            bitOffset -= 8;
+            if (byteOffset >= len) {
+                ESP_LOGW(TAG, "Parameter 0x%04x has bit offset %d but message only has %d bytes", paramDesc->parameterId, paramDesc->sensorType, len);
+                return;
             }
-            bool value = !!(data[byteOffset] & (1 << bitOffset));
-            binarySensor->publish_state(value);
-        } else {
-            ESP_LOGW(TAG, "Sensor type %d NYI for binary sensor", paramDesc->sensorType);
         }
-
-    } else if(switch_) {
-        if (paramDesc->sensorType == TAG_NACHT_AUTO_SELECT) {
-            ESP_LOGD(TAG, "Found tag/nacht/auto select length %d: 0x%02x (0:Nacht, 1:Tag, 2: Auto) 0x%02x", len, data[0], data[1] );
-            if(data[0] == 0) { // nacht
-                switch_->publish_state(false);
-            } else if(data[0] == 1 || data[0] == 2) { // tag, auto
-                switch_->publish_state(true);
-            } else {
-                ESP_LOGW(TAG, "Invalid value for tag/nacht/auto select: %d 0x%02x 0x%02x ...", len, data[0], data[1]);
-            }
+        bool value = !!(data[byteOffset] & (1 << bitOffset));
+        handleReceivedUnsignedValue(paramDesc->sensorTypeParam, value);
+    } else if (paramDesc->sensorType == TAG_NACHT_AUTO_SELECT) {
+        ESP_LOGD(TAG, "Found tag/nacht/auto select length %d: 0x%02x (0:Nacht, 1:Tag, 2: Auto) 0x%02x", len, data[0], data[1] );
+        if(data[0] == 0) { // nacht
+            handleReceivedUnsignedValue(paramDesc->sensorTypeParam, 0);
+        } else if(data[0] == 1 || data[0] == 2) { // tag, auto
+            handleReceivedUnsignedValue(paramDesc->sensorTypeParam, 1);
         } else {
-            ESP_LOGW(TAG, "Sensor type %d NYI for switch", paramDesc->sensorType);
+            ESP_LOGW(TAG, "Invalid value for tag/nacht/auto select: %d 0x%02x 0x%02x ...", len, data[0], data[1]);
         }
-    } else if(number) {
-        if (paramDesc->sensorType == UNSIGNED_INT) {
-            uint32_t value = parseUnsignedInteger(data, len);
-            number->publish_state(value);
-        } else if (paramDesc->sensorType == SIGNED_INT) {
-            int32_t value = parseSignedInteger(data, len);
-            number->publish_state(value);
-        } else if(paramDesc->sensorType == BYTE_AT_OFFSET) {
-            if (paramDesc->sensorTypeParam < len) {
-                uint8_t value = data[paramDesc->sensorTypeParam];
-                number->publish_state(value);
-            } else {
-                ESP_LOGE(TAG, "Offset for sensor type BYTE_AT_OFFSET %d > data len %d", paramDesc->sensorTypeParam, len);
-            }
+    } else if(paramDesc->sensorType == BYTE_AT_OFFSET) {
+        if (paramDesc->sensorTypeParam < len) {
+            uint8_t value = data[paramDesc->sensorTypeParam];
+            handleReceivedUnsignedValue(paramDesc->sensorTypeParam, value);
         } else {
-            ESP_LOGE(TAG, "Sensor type %d NYI for number", paramDesc->sensorType);
+            ESP_LOGE(TAG, "Offset for sensor type BYTE_AT_OFFSET %d > data len %d", paramDesc->sensorTypeParam, len);
         }
-    } else if(select) {
-        if(paramDesc->sensorType == BYTE_AT_OFFSET) {
-            if (paramDesc->sensorTypeParam < len) {
-                uint8_t value = data[paramDesc->sensorTypeParam];
-                select->handleReceivedValue(value);
-            } else {
-                ESP_LOGE(TAG, "Offset for sensor type BYTE_AT_OFFSET %d > data len %d", paramDesc->sensorTypeParam, len);
-            }
-        } else {
-            ESP_LOGE(TAG, "Sensor type %d NYI for select", paramDesc->sensorType);
-        }
-    } else if(assembler) {
-        if (paramDesc->sensorType == MULTI_PARAMETER_UNSIGNED_INTEGER) {
-            ESP_LOGD(TAG, "Parameter: 0x%04x type param %d Len %d, data: %d %d", paramDesc->parameterId, paramDesc->sensorTypeParam, len, data[0], data[1]);
-            uint32_t value = parseUnsignedInteger(data, len);
-            assembler->handleReceivedValue(paramDesc->sensorTypeParam, value);
-        } else {
-            ESP_LOGE(TAG, "Sensor type %d NYI for MultiParameterUnsignedIntegerAssembler", paramDesc->sensorType);
-        }
-
+    } else if (paramDesc->sensorType == MULTI_PARAMETER_UNSIGNED_INTEGER) {
+        ESP_LOGD(TAG, "Parameter: 0x%04x type param %d Len %d, data: %d %d", paramDesc->parameterId, paramDesc->sensorTypeParam, len, data[0], data[1]);
+        uint32_t value = parseUnsignedInteger(data, len);
+        handleReceivedUnsignedValue(paramDesc->sensorTypeParam, value);
     } else {
-        ESP_LOGE(TAG, "No sensor, binary sensor, switch or number defined");
+        ESP_LOGW(TAG, "Sensor type %d NYI", paramDesc->sensorType);
     }
 }
 
 void BuderusValueHandler::loop()
 {
-  if(number) {
-      number->loop();
-  }
+    if(communicationComponent) {
+        communicationComponent->loop();
+    }
 }
 
-BuderusParamSwitch::BuderusParamSwitch():
-    writer(nullptr)
+void BuderusValueHandler::handleReceivedSignedValue(uint16_t sensorTypeParam, int32_t value)
 {
-}
-
-void BuderusParamSwitch::setupWriting(Writer3964R *writer, Buderus_R2017_ParameterId parameterId, SensorType sensorType)
-{
-    this->writer = writer;
-    this->parameterId = parameterId;
-    this->sensorType = sensorType;
-}
-
-void BuderusParamSwitch::write_state(bool state)
-{
-    const uint8_t keep = 0x65;
-    const uint8_t data_type_warm_water = 0x0c;
-
-    if (parameterId == CFG_WW_Aufbereitung) {
-        const uint8_t automatic = 2;
-        const uint8_t day = 1;
-        const uint8_t night = 0;
-
-        const uint8_t value = state ? automatic : night;
-
-        const uint8_t message[] = { data_type_warm_water, 0x0E, value, keep, keep, keep, keep, keep};
-        writer->enqueueTelegram(message, 8);
-        publish_state(state); // confirm for now without waiting for an ack from the heater
-
+    if (sensor) {
+        sensor->publish_state(value);
+    } else if(binarySensor) {
+        binarySensor->publish_state(value);
+    } else if (communicationComponent) {
+        communicationComponent->handleReceivedSignedValue(sensorTypeParam, value);
     } else {
-        ESP_LOGE(TAG, "No write configuration for parameter 0x%04x found", parameterId);
+        ESP_LOGW(TAG, "No handler for signed values found at %d", paramDesc->sensorType);
     }
 }
 
-BuderusParamNumber::BuderusParamNumber():
-    writer(nullptr),
-    hasPendingWriteRequest(false)
+void BuderusValueHandler::handleReceivedUnsignedValue(uint16_t sensorTypeParam, uint32_t value)
 {
-}
-
-void BuderusParamNumber::setupWriting(Writer3964R *writer, Buderus_R2017_ParameterId parameterId, SensorType sensorType)
-{
-    this->writer = writer;
-    this->parameterId = parameterId;
-    this->sensorType = sensorType;
-}
-
-void BuderusParamNumber::control(float value)
-{
-    if(parameterId == CFG_WW_Temperatur) {
-        float target_temperature = value;
-        if (target_temperature < 30) {
-            target_temperature = 30;
-        } else if (target_temperature > 60) {
-            target_temperature = 60;
-        }
-
-        // do not write directly, but instead wait until the value does not change for some time and write then
-        // this is to avoid writing to the heater each time the user clicks on the up arrow
-        this->hasPendingWriteRequest = true;
-        this->pendingWriteValue = target_temperature;
-        this->lastWriteRequest = millis();
-
-    } else if(parameterId == CFG_HK1_Auslegungstemperatur) {
-        float target_temperature = value;
-        if (target_temperature < 30) {
-            target_temperature = 30;
-        } else if (target_temperature > 90) {
-            target_temperature = 90;
-        }
-
-        // do not write directly, but instead wait until the value does not change for some time and write then
-        // this is to avoid writing to the heater each time the user clicks on the up arrow
-        this->hasPendingWriteRequest = true;
-        this->pendingWriteValue = target_temperature;
-        this->lastWriteRequest = millis();
+    if (sensor) {
+        sensor->publish_state(value);
+    } else if(binarySensor) {
+        binarySensor->publish_state(value);
+    } else if (communicationComponent) {
+        communicationComponent->handleReceivedUnsignedValue(sensorTypeParam, value);
     } else {
-        ESP_LOGE(TAG, "No write configuration for number parameter 0x%04x found", parameterId);
-    }
-}
-
-void BuderusParamNumber::loop()
-{
-    uint32_t writeConsolidationPeriod = 1000;
-    const uint8_t keep = 0x65;
-    const uint8_t data_type_warm_water = 0x0c;
-    const uint8_t data_type_heating_circuit_1 = 0x07;
-
-    if (this->hasPendingWriteRequest) {
-        uint32_t now = millis();
-        if (now - lastWriteRequest > writeConsolidationPeriod) {
-
-            if (parameterId == CFG_WW_Temperatur) {
-              const uint8_t message[] = { data_type_warm_water, 0x07, keep, keep, keep, (uint8_t)this->pendingWriteValue, keep, keep};
-                if(writer->enqueueTelegram(message, 8)) {
-                    this->hasPendingWriteRequest = false;
-                    publish_state(this->pendingWriteValue);
-
-                }
-            } else if(parameterId == CFG_HK1_Auslegungstemperatur) {
-                const uint8_t message[] = { data_type_heating_circuit_1, 0x0e, keep, keep, keep, keep, (uint8_t)this->pendingWriteValue, keep};
-                  if(writer->enqueueTelegram(message, 8)) {
-                      this->hasPendingWriteRequest = false;
-                      publish_state(this->pendingWriteValue);
-
-                  }
-            } else {
-                ESP_LOGE(TAG, "No support for writing parameter 0x%04x", parameterId);
-                this->hasPendingWriteRequest = false;
-
-            }
-        }
-    }
-}
-
-MultiParameterUnsignedIntegerAssembler::MultiParameterUnsignedIntegerAssembler(esphome::sensor::Sensor *targetSensor):
-    sensor(targetSensor),
-    component_known{false, false, false}
-{
-
-}
-
-void MultiParameterUnsignedIntegerAssembler::handleReceivedValue(uint16_t sensorTypeParam, uint8_t value)
-{
-    ESP_LOGD(TAG, "Received value for st param %d: %d", sensorTypeParam, value);
-
-    int valueIndex = sensorTypeParam & 0x0f;
-    if (valueIndex > 2) {
-        ESP_LOGE(TAG, "Invalid sensor type param: %d", sensorTypeParam);
-        return;
-    }
-    components[valueIndex] = value;
-    component_known[valueIndex] = true;
-    if (valueIndex != 0) { // only update the sensor value on lsb updates to avoid jumps
-       return;
+        ESP_LOGW(TAG, "No handler for unsigned values found at %d", paramDesc->sensorType);
     }
 
-
-    bool all_components_known = component_known[0] && component_known[1] && component_known[2];
-    if (all_components_known) {
-       uint32_t result = (((components[2] << 8) + components[1]) << 8 )  + components[0];
-       ESP_LOGD(TAG, "Assembling %d %d %d to %d", components[0], components[1], components[2], result);
-       sensor->publish_state(result);
-    }
 }
 
-BuderusParamSelect::BuderusParamSelect():
-    writer(nullptr)
+void BuderusValueHandler::handleReceivedFloatValue(uint16_t sensorTypeParam, float value)
 {
-
-}
-
-void BuderusParamSelect::setupWriting(Writer3964R *writer, Buderus_R2017_ParameterId parameterId, SensorType sensorType, uint16_t sensorTypeParam)
-{
-    this->writer = writer;
-    this->parameterId = parameterId;
-    this->sensorType = sensorType;
-    this->sensorTypeParam = sensorTypeParam;
-}
-
-void BuderusParamSelect::setSelectMappings(std::vector<uint8_t> mappings)
-{
-    this->mappings = std::move(mappings);
-}
-
-void BuderusParamSelect::handleReceivedValue(uint8_t value)
-{
-    auto it = std::find(mappings.cbegin(), mappings.cend(), value);
-    if (it == mappings.end()) {
-        ESP_LOGE(TAG, "Invalid value %u received for select of parameter 0x%04x", value, parameterId);
-        return;
-    }
-    size_t mapping_idx = std::distance(mappings.cbegin(), it);
-    auto selectValue = this->at(mapping_idx);
-    this->publish_state(selectValue.value());
-}
-
-void BuderusParamSelect::control(const std::string &value) {
-
-    const uint8_t keep = 0x65;
-    const uint8_t data_type_warm_water = 0x0c;
-    const uint8_t data_type_heating_circuit_1 = 0x07;
-
-    auto idx = this->index_of(value);
-
-    if (idx.has_value()) {
-        uint8_t numericValue= this->mappings.at(idx.value());
-        if (parameterId == CFG_HK1_Betriebsart && this->sensorTypeParam == 4) {
-            if (numericValue > 2) {
-                ESP_LOGE(TAG, "Invalid select value for parameter %d received: 0x%04x", parameterId, numericValue);
-                return;
-            }
-
-            const uint8_t message[] = { data_type_heating_circuit_1, 0x00, keep, keep, keep, keep, (uint8_t )numericValue, keep};
-            writer->enqueueTelegram(message, 8);
-            publish_state(value); // confirm for now without waiting for an ack from the heater
-        } else if (parameterId == CFG_WW_Aufbereitung && this->sensorTypeParam == 0) {
-            if (numericValue > 2) {
-                ESP_LOGE(TAG, "Invalid select value for parameter %d received: 0x%04x", parameterId, numericValue);
-                return;
-            }
-
-            const uint8_t message[] = { data_type_warm_water, 0x0e, (uint8_t)numericValue, keep, keep, keep, keep, keep};
-            writer->enqueueTelegram(message, 8);
-            publish_state(value); // confirm for now without waiting for an ack from the heater
-        } else {
-            ESP_LOGE(TAG, "No write configuration for parameter 0x%04x found", parameterId);
-        }
+    if (sensor) {
+        sensor->publish_state(value);
+    } else if(binarySensor) {
+        binarySensor->publish_state(value);
+    } else if (communicationComponent) {
+        communicationComponent->handleReceivedFloatValue(sensorTypeParam, value);
     } else {
-        ESP_LOGE(TAG, "No mapping for select value %s found", value.c_str());
+        ESP_LOGW(TAG, "No handler for float values found at %d", paramDesc->sensorType);
     }
 }
+
 
 } // namespace
 } // namespace
